benchmarks/Text2SpeechCensoringWorkflow/ Text2SpeechCensoringWorkflow_Conversion/requirements.txt
# Function dependencies, for example:
# package>=version
pydub>=0.23.1
google-cloud-storage==1.43.0
Flask==2.1.0
itsdangerous==2.0.1
google-cloud-datastore==1.13.2
google-cloud-pubsub>=0.28.1